feat: add data wrangling and model build and eval
This is the inital commit. Add jupter notebook, data files and
script files to respository. Script files perform data wrangling and
builds the classification model that assigns text message to the right
emergency relief category. Running script extracts and processes data,
trains data on the model, and evaluates and saves the model to disk.